🐛 fix graph values displaying incorrect data

+++ web-laravel-monetary-budget/app/Models/Budget.php
@@ -18,7 +18,7 @@ class Budget extends Model
     ];
 
     protected $casts = [
-        'amount' => 'integer',
+        'amount' => 'double',
     ];
 
     protected $appends = [
